+++ packages/labs/examples/index.ts
@@ -12,5 +12,4 @@ export * from "./selectExample";
 export * from "./suggestExample";
 export * from "./tagInputExample";
 export * from "./tooltip2Example";
-export * from "./timezoneSelectBasicExample";
-export * from "./timezoneSelectExtendedExample";
+export * from "./timezoneSelectExample";
