+++ packages/labs/src/components/timezone-select/timezone-select.md
@@ -1,8 +1,8 @@
 @# Timezone Select
 
-`TimezoneSelect` allows the user select from a list of timezones.
+`TimezoneSelect` allows the user to select from a list of timezones.
 
-[Moment Timezone](http://momentjs.com/timezone/) is used internally,
+[Moment Timezone](http://momentjs.com/timezone/) is used internally
 for the list of available timezones and timezone metadata.
 
 This component can be used in controlled or uncontrolled mode.
@@ -21,18 +21,17 @@ using the most populous location for each offset.
 Moment Timezone uses a similar heuristic for
 [guessing](http://momentjs.com/timezone/docs/#/using-timezones/guessing-user-timezone/) the user's timezone.
 
-Note about `defaultToLocalTimezone` and `showLocalTimezone`:
-We cannot guarantee that we'll get the correct local timezone in all browsers.
-In supported browsers, the Internationalization API is used.
-In other browsers, Date methods and a population heuristic are used.
-See [Moment Timezone's documentation](https://momentjs.com/timezone/docs/#/using-timezones/guessing-user-timezone/)
-for more information.
+<div class="pt-callout pt-intent-warning pt-icon-warning-sign">
+    <h5>Local timezone detection</h5>
+    We detect the local timezone when the `defaultToLocalTimezone` and `showLocalTimezone` props are used.
+    We cannot guarantee that we'll get the correct local timezone in all browsers.
+    In supported browsers, the [i18n API](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DateTimeFormat/resolvedOptions) is used.
+    In other browsers, `Date` methods and a population heuristic are used.
+    See [Moment Timezone's documentation](https://momentjs.com/timezone/docs/#/using-timezones/guessing-user-timezone/)
+    for more information.
+</div>
 
-@## Basic example
-
-@reactExample TimezoneSelectBasicExample
-
-@## Basic usage
+@reactExample TimezoneSelectExample
 
 ```tsx
 import { TimezoneSelect } from "@blueprintjs/core";
@@ -42,7 +41,9 @@ export interface ITimezoneExampleState {
 }
 
 export class TimezoneExample extends React.PureComponent<{}, ITimezoneExampleState> {
-    public state: ITimezoneExampleState = {};
+    public state: ITimezoneExampleState = {
+        timezone: "",
+    };
 
     public render() {
         return (
@@ -53,18 +54,10 @@ export class TimezoneExample extends React.PureComponent<{}, ITimezoneExampleSta
         );
     }
 
-    private handleTimezoneChange = (timezone: string) => {
-        this.setState({ timezone });
-    }
+    private handleTimezoneChange = (timezone: string) => this.setState({ timezone });
 }
 ```
-
-@## Extended example
-
-@reactExample TimezoneSelectExtendedExample
 
 @## JavaScript API
 
 @interface ITimezoneSelectProps
-
-@interface ITimezoneSelectTargetProps
